refactor: clean up organizations crud implementation
