+++ examples/app/organizations/page.tsx
@@ -9,19 +9,42 @@ const CrudModule = dynamic(
 
 import { CustomSelectWidget } from '@concepta/react-material-ui/dist/styles/CustomWidgets';
 
-const getMaskedPhone = (text: string) => {
-  const numericInput = text.replace(/\D/g, '');
-  const truncatedInput = numericInput.slice(0, 10);
-  const phoneNumberFormat = /^(\d{3})?(\d{3})?(\d{0,4})?$/;
-  const formattedNumber = truncatedInput.replace(
-    phoneNumberFormat,
-    (_, p1, p2, p3) => {
-      const formattedGroups = [p1 && `(${p1})`, p2 && ` ${p2}`, p3 && `-${p3}`];
-      return formattedGroups.filter(Boolean).join('');
-    },
-  );
+import { getMaskedPhone } from '@/utils/mask';
 
-  return formattedNumber;
+const defaultFormSchema = {
+  type: 'object',
+  properties: {
+    name: { type: 'string', title: 'Name' },
+    address: { type: 'string', title: 'Address' },
+    city: { type: 'string', title: 'City' },
+    state: { type: 'string', title: 'State' },
+    phone: { type: 'string', title: 'Phone' },
+    owner: {
+      type: 'number',
+      title: 'Owner',
+      oneOf: [
+        { const: 1, title: 'John Doe' },
+        { const: 2, title: 'Jane Doe' },
+      ],
+    },
+  },
+};
+
+const defaultFormUiSchema = {
+  city: {
+    'ui:gridColumns': 6,
+  },
+  state: {
+    'ui:gridColumns': 6,
+  },
+  phone: {
+    'ui:gridColumns': 6,
+    'ui:formatter': (value: string) => getMaskedPhone(value),
+  },
+  owner: {
+    'ui:widget': CustomSelectWidget,
+    'ui:gridColumns': 6,
+  },
 };
 
 const Organizations = () => {
@@ -64,112 +87,21 @@ const Organizations = () => {
       }}
       createFormProps={{
         formSchema: {
-          type: 'object',
-          properties: {
-            name: { type: 'string', title: 'Name' },
-            address: { type: 'string', title: 'Address' },
-            city: { type: 'string', title: 'City' },
-            state: { type: 'string', title: 'State' },
-            phone: { type: 'string', title: 'Phone' },
-            owner: {
-              type: 'number',
-              title: 'Owner',
-              oneOf: [
-                { const: 1, title: 'John Doe' },
-                { const: 2, title: 'Jane Doe' },
-              ],
-            },
-          },
+          ...defaultFormSchema,
+          required: ['name', 'address', 'city', 'state', 'phone', 'owner'],
         },
-        formUiSchema: {
-          city: {
-            'ui:gridColumns': 6,
-          },
-          state: {
-            'ui:gridColumns': 6,
-          },
-          phone: {
-            'ui:gridColumns': 6,
-            'ui:formatter': (value: string) => getMaskedPhone(value),
-          },
-          owner: {
-            'ui:widget': CustomSelectWidget,
-            'ui:gridColumns': 6,
-          },
-        },
+        formUiSchema: defaultFormUiSchema,
       }}
       editFormProps={{
-        formSchema: {
-          type: 'object',
-          properties: {
-            name: { type: 'string', title: 'Name' },
-            address: { type: 'string', title: 'Address' },
-            city: { type: 'string', title: 'City' },
-            state: { type: 'string', title: 'State' },
-            phone: { type: 'string', title: 'Phone' },
-            owner: {
-              type: 'number',
-              title: 'Owner',
-              oneOf: [
-                { const: 1, title: 'John Doe' },
-                { const: 2, title: 'Jane Doe' },
-              ],
-            },
-          },
-        },
-        formUiSchema: {
-          city: {
-            'ui:gridColumns': 6,
-          },
-          state: {
-            'ui:gridColumns': 6,
-          },
-          phone: {
-            'ui:gridColumns': 6,
-            'ui:formatter': (value: string) => getMaskedPhone(value),
-          },
-          owner: {
-            'ui:widget': CustomSelectWidget,
-            'ui:gridColumns': 6,
-          },
-        },
+        formSchema: defaultFormSchema,
+        formUiSchema: defaultFormUiSchema,
       }}
       detailsFormProps={{
         formSchema: {
-          type: 'object',
-          properties: {
-            name: { type: 'string', title: 'Name' },
-            address: { type: 'string', title: 'Address' },
-            city: { type: 'string', title: 'City' },
-            state: { type: 'string', title: 'State' },
-            phone: { type: 'string', title: 'Phone' },
-            owner: {
-              type: 'number',
-              title: 'Owner',
-              oneOf: [
-                { const: 1, title: 'John Doe' },
-                { const: 2, title: 'Jane Doe' },
-              ],
-            },
-          },
+          ...defaultFormSchema,
           required: ['name', 'address', 'city', 'state', 'phone', 'owner'],
         },
-        formUiSchema: {
-          city: {
-            'ui:gridColumns': 6,
-          },
-          state: {
-            'ui:gridColumns': 6,
-          },
-          phone: {
-            'ui:gridColumns': 6,
-            'ui:formatter': (value: string) => getMaskedPhone(value),
-          },
-          owner: {
-            'ui:widget': CustomSelectWidget,
-            'ui:gridColumns': 6,
-          },
-        },
+        formUiSchema: defaultFormUiSchema,
       }}
     />
   );
